refactor(fixture): fixture directory로  mocks data 이동
