+++ src/assets/data/index.ts
@@ -2,4 +2,3 @@ export { default as emptyCategories } from "./emptyCategories";
 export { default as emptyClassInfo } from "./emptyClassInfo";
 export { default as emptyClassLevel } from "./emptyClassLevel";
 export { default as emptyTimetable } from "./emptyTimetable";
-export { default as emptyPostList } from "./emptyPostList";
