+++ src/fixture/index.ts
@@ -1,0 +1,2 @@
+export { default as post } from "./post";
+export { default as postList } from "./postList";
