+++ src/page/forum-post/layouts/Post.tsx
@@ -3,8 +3,8 @@ import styled from "styled-components";
 import { useRouter } from "next/navigation";
 import { ROUTER } from "@/constants";
 import { color } from "@/styles";
-import emptyPost from "@/assets/data/emptyPost";
 import { IPost } from "@/interfaces";
+import { post as emptyPost } from "@/fixture";
 import PostBody from "./PostBody";
 import PostHead from "./PostHead";
 import { usePostQuery } from "../services/query.service";
